Board(盤面)クラスを実装し、ボタンの処理をそっちに移動 ゲームコントローラーを一先ず空に シングルトン基底クラスも実装内容を更新
