+++ Assets/Scenes/SampleScene.unity
@@ -260,6 +260,8 @@ GameObject:
   m_Component:
   - component: {fileID: 1505447037}
   - component: {fileID: 1505447036}
+  - component: {fileID: 1505447039}
+  - component: {fileID: 1505447038}
   m_Layer: 0
   m_Name: BaseObject
   m_TagString: Untagged
@@ -273,7 +275,7 @@ MonoBehaviour:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 1505447035}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: cd5eeca9ae5786c4eb29712664a7f1c8, type: 3}
   m_Name: 
@@ -293,3 +295,27 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1505447038
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1505447035}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 44fbcbe3f2a96844dbead79a12044a09, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _button_prefab: {fileID: 1874334094572816, guid: 76f3d2e8db4237c4c868618b1b8321c0,
+    type: 2}
+--- !u!114 &1505447039
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1505447035}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 690363226284c7547b1b47ab38821872, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
