+++ Assets/Scripts/InputManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 690363226284c7547b1b47ab38821872
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
