+++ Assets/Scripts/SingletonMonoBehavior.cs
@@ -10,6 +10,11 @@ public class SingletonMonoBehavior<T> : MonoBehaviour where T : SingletonMonoBeh
 			if(_instance == null)
 			{
 				_instance = FindObjectOfType <T>();
+
+				if (_instance == null)
+				{
+					_instance = new GameObject("Manager").AddComponent<T>();
+				}
 			}
 			return _instance;
 		}
